Rename extensions/loom/session-bootstrap.ts to session-lifecycle.ts. The file already owns both startup AND shutdown handlers (session_start, session_before_compact, session_shutdown) -- "bootstrap" only describes the startup half. The new name matches what's actually in the file and lines up with where session-end protocol logic will land next. Pure rename: no content change in the .ts file. Updated import in extensions/loom/index.ts and references in CLAUDE.md and docs/architecture.md.
